shiboken: Make it possible to extend namespaces
Qt has some modules, where functions and classes are added to
namespaces of other modules. For example, QtGui adds the free
functions mightBeRichText() and convertFromPlainText() to the "Qt"
namespace and Qt3DQuick adds a namespace "Quick" to the namespace
"Qt3DCore" started in Qt3DCore. Shiboken was unable to handle this
since the name of the index constant was derived from the namespace
name and would thus clash. Also, all code model items from the base
module would be re-generated into the extended namespace.

To fix this:

- No longer join namespace code model items in clang builder.

- Search for pre-existing namespaces in traverseNamespace() before
  creating a new one, continuing populating it.

- Add a "files" attribute taking a regular expression to match header
  names to the namespace type entry, making it possible to restrict
  code model items by files when extending namespaces (as otherwise,
  all base items are again added).

- Add "extends" attribute specifying the package name of the namespace
  to extend, which will be used as base type in CPython.

- Disambiguate the SBK indexes of namespaces and the init_ functions
  by adding the module name.

Change-Id: Ib552e878911942fa4d52d2bb0c6695e5b2c4c341
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>