shiboken: Fix out of bounds string access
Fix warning:
(shiboken) Using QCharRef with an index pointing outside the valid range of a QString. The corresponding behavior is deprecated, and will be changed in a future version of Qt.

introduced by qtbase/c2d2757bccc68e1b981df059786c2e76f2969530 (5.14).
This happens when endPos is modified by the prior call to parseType().

Change-Id: I3b81e674c0aa118bd003aa25d8c61d89d772366f
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>