shiboken: Store the unqualified entry name in TypeEntry
With the previous change adding a parent pointer, this is working
towards building the target lang name by walking up the hierarchy,
prepending the names, making it possible to exclude namespaces.

Pass the unqualified name from the XML parser and build
the qualified name in the TypeEntry constructor.

For this to work, a new ConstantValueTypeEntry is added replacing the
abuse of EnumValueTypeEntry for nontype-template parameters.

As a side effect, it is no longer possible to nest types
by qualifying with "::" in XML:

<object-type name="Class"/>
<enum-type name="Class::Enum"/>

This needs to  be fixed in the type system files.

[ChangeLog][shiboken] As a result of a code cleanup, it is
no longer possible to nest types by by qualifying with "::" in
the type system files. The elements need to be properly nested.

Task-number: PYSIDE-990
Task-number: PYSIDE-1074
Change-Id: I8a2f93c40d59167b0ba205ef3ff3b325d242c3d3
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>