shiboken: Fix passing of the parent type to smart pointer init functions
The code was assuming that smart pointers instances live in the global
namespace and passed the module as parent, which does not work in case
of std::shared_ptr.

Factor out the code writing the init function call and use the same
code for classes and smart pointer instances.

Task-number: PYSIDE-454
Change-Id: Iffe5ace31d734dd19ca784841344c50248952342
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>